refactor: extract EpisodeSearchPanel and EpisodeHistoryPanel
